+++ CHANGELOG.rst
@@ -1,6 +1,17 @@
 Change Log
 ==========
 
+v22.9.2
+--------
+
+Main changes:
+
+- Update CLI commands (snb-parse, analyse, plot and groundstate can all now be run with no arguments within a defect folder)
+- Update custom font
+- Update groundstate() tests
+- Update plotting
+
+
 v22.9.1
 --------